requirements.txt
biopython
pandas
scikit-learn
gensim
matplotlib
nltk
numpy
seaborn